requirements.txt
baostock==0.8.9
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
curl_cffi==0.15.0
decorator==5.2.1
et_xmlfile==2.0.0
frozendict==2.4.7
futu_api==10.2.6218
html5lib==1.1
idna==3.11
Jinja2==3.1.6
jsonpath==0.82.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mini-racer==0.14.1
multitasking==0.0.12
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
patsy==1.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.6
prettytable==3.17.0
protobuf==3.20.3
pycparser==3.0
pycryptodome==3.23.0
pyecharts==2.1.0
Pygments==2.20.0
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
rich==14.3.3
scipy==1.17.1
simplejson==3.20.2
six==1.17.0
soupsieve==2.8.3
statsmodels==0.14.6
tabulate==0.10.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
xlrd==2.0.2
yfinance==1.2.1
